requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
igdb-api-v4==0.3.3
protobuf==5.29.3
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.3.0
